Include <config.h> in C files that include mono headers
Any translation unit that includes a Mono header should include <config.h>
first to ensure that all of our configuration defines are available.

In particular, config.h defines MONO_INSIDE_RUNTIME which, in mono-publib.h causes
MONO_RT_EXTERNAL_ONLY to be defined to an attribute that causes a compilation
error if you use a function that is internally deprecated for the runtime.

Also change a couple of places that used non-_checked functions.