Rename Mono.Posix.* -> Mono.Unix.*
Fixes: https://github.com/mono/mono.posix/issues/1
Fixes: https://github.com/mono/mono.posix/issues/2

Managed assembly renamed to `Mono.Unix.dll` and the native helper
library renamed to `libMono.Unix.{so,dylib,a}`. This is done to reflect
the fact that we're breaking API and ABI.

Move tests to `tests/` and rename the test assembly to
`Mono.Unix.Test.dll`

Note: tests appear to hang after the rename while they used to run to
completion before.